class_to_parse_lambda, move to Model (before Import) due to options check
