+++ spec/csv_row_model/concerns/model/attributes_spec.rb
@@ -106,6 +106,57 @@ describe CsvRowModel::Model::Attributes do
           end
         end
       end
+
+      describe "::class_to_parse_lambda" do
+        subject { klass.class_to_parse_lambda }
+
+        it "returns the CLASS_TO_PARSE_LAMBDA" do
+          expect(subject).to eql CsvRowModel::Import::Attributes::CLASS_TO_PARSE_LAMBDA
+        end
+
+        context "with module overriding it" do
+          before do
+            module Overrides
+              extend ActiveSupport::Concern
+              include CsvRowModel::Import
+
+              module ClassMethods
+                def class_to_parse_lambda; {} end
+              end
+            end
+
+            klass.send(:include, Overrides)
+          end
+
+          it "overrides properly" do
+            expect(subject).to eql({})
+          end
+        end
+      end
+
+      describe "::custom_check_options" do
+        subject { klass.custom_check_options(options) }
+
+        context "with invalid :type Option" do
+          let(:options) { { type: Object } }
+
+          it "raises exception" do
+            expect { subject }.to raise_error(":type must be Boolean, String, Integer, Float, DateTime, Date")
+          end
+
+          context "with ::class_to_parse_lambda overwritten" do
+            before do
+              klass.class_eval do
+                def self.class_to_parse_lambda; super.merge(Hash => ->(s) { JSON.parse(s) }) end
+              end
+            end
+
+            it "raises a new type of exception" do
+              expect { subject }.to raise_error(":type must be Boolean, String, Integer, Float, DateTime, Date, Hash")
+            end
+          end
+        end
+      end
     end
   end
 end
